PythonSDK_Turnlamp/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile ./builds/demos/templates/common/pyproject.toml -o ./builds/demos/templates/common/requirements.txt
allure-pytest==2.15.0
    # via openta-pytest
allure-python-commons==2.15.0
    # via allure-pytest
attrs==25.3.0
    # via
    #   allure-python-commons
    #   jsonschema
    #   referencing
certifi==2025.8.3
    # via requests
cffi==1.17.1
    # via clr-loader
charset-normalizer==3.4.3
    # via requests
clr-loader==0.2.7.post0
    # via pythonnet
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.59.1
    # via matplotlib
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
jsonschema==4.25.0
    # via openta
jsonschema-specifications==2025.4.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.5
    # via pythonsdk (./builds/demos/templates/common/pyproject.toml)
numpy==2.3.2
    # via
    #   contourpy
    #   matplotlib
    #   openta
    #   pandas
wheels/openta-0.1.1b3-py3-none-any.whl
    # via
    #   pythonsdk (./builds/demos/templates/common/pyproject.toml)
    #   openta-pytest
wheels/openta_pytest-0.1.1b3-py3-none-any.whl
    # via pythonsdk (./builds/demos/templates/common/pyproject.toml)
packaging==25.0
    # via
    #   matplotlib
    #   pytest
pandas==2.3.1
    # via openta
pillow==11.3.0
    # via matplotlib
pluggy==1.6.0
    # via
    #   allure-python-commons
    #   pytest
pycparser==2.22
    # via cffi
pygments==2.19.2
    # via pytest
pyparsing==3.2.3
    # via matplotlib
pytest==8.4.1
    # via
    #   pythonsdk (./builds/demos/templates/common/pyproject.toml)
    #   allure-pytest
    #   openta-pytest
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pythonnet==3.0.5
    # via openta
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via openta
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via openta
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
typing-extensions==4.14.1
    # via referencing
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
